requirements.txt
streamlit==1.40.0
langchain==0.3.14
langchain-openai==0.2.14
langchain-experimental==0.3.4
pandas==2.2.3
python-dotenv==1.0.1
tabulate==0.9.0
